+++ changelog.md
@@ -1,4 +1,33 @@
 # CS-30-Final-Project
 <br><br>
-
+x.y.z  
+x: represents the coder ID  
+- Amir uses code 0  
+- Leo uses code 1  
+y: the version number  
+z: commit number  
+example:  
+0.1.1 means amir comitted his first commit to version 1.  
 ## Change Log
+> 12.20  
+
+1.1.# 
+- finish *player* folder
+    + bodyParts.json
+    + player.py
+    + playingRecord.json
+- finish `main.py` in main folder
+    + can now import all other four mini-games
+    + creat four objects inside `Game` class
+    + can run the four games inside folder
+    + import `player.py` from *player* folder
+    + creat a player object to read player information, store player information and store playing record
+>12.20  
+
+- finish game 2
+    > 12.21  
+    + add the rules of the game and stored it as a `.txt` file (`rules..txt` in *game 2* folder)
+    + add some animation(print rules word by word, add screen flickering when the player lost and loses a body part)
+    + debug game 2 
+    + `main.py` can now read results from `main.py` in *game 2* folder and update score, game results and lost body parts in `playingRecord.json` in *player* folder
+